+++ demisto_sdk/commands/common/hook_validations/correlation_rule.py
@@ -5,14 +5,16 @@ This module is designed to validate the correctness of generic definition entiti
 
 from ruamel.yaml.comments import CommentedSeq
 
+from demisto_sdk.commands.common.constants import (
+    FILETYPE_TO_DEFAULT_FROMVERSION,
+    FileType,
+)
 from demisto_sdk.commands.common.errors import Errors
 from demisto_sdk.commands.common.hook_validations.base_validator import error_codes
 from demisto_sdk.commands.common.hook_validations.content_entity_validator import (
     ContentEntityValidator,
 )
 from demisto_sdk.commands.common.logger import logger
-
-FROMVERSION = "6.10.0"
 
 
 class CorrelationRuleValidator(ContentEntityValidator):
@@ -30,7 +32,9 @@ class CorrelationRuleValidator(ContentEntityValidator):
             structure_validator,
             ignored_errors=ignored_errors,
             json_file_path=json_file_path,
-            oldest_supported_version=FROMVERSION,
+            oldest_supported_version=FILETYPE_TO_DEFAULT_FROMVERSION[
+                FileType.CORRELATION_RULE
+            ],
         )
         self._is_valid = True